--- VideoApp/VideoApp/NewUI/Screens/Room/StreamConfig.swift
@@ -1,10 +1,0 @@
-//
-//  Copyright (C) 2021 Twilio, Inc.
-//
-
-import Foundation
-
-struct StreamConfig {
-    let streamName: String
-    let userIdentity: String
-}
