+++ VideoApp/VideoApp/NewUI/Views/Core/DisplayNameFactory.swift
@@ -4,6 +4,7 @@
 
 import Foundation
 
+// TODO: Rename to something "user"
 class DisplayNameFactory {
     func makeDisplayName(identity: String, isYou: Bool = false) -> String {
         isYou ? "You" : identity
